+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Translator features with MSSQL support.
+- Dialect Enum.
+
+### Changed
+- Dataset sql(&self, query: &str) method changed to relation(&self, query: &str, dialect: Option<Dialect>).
+- Added optional dialect in the signature of Dataset's from_queries method.
+- Relation render() method changed to to_query(dialect: Option<Dialect>).
+- Relation parse(query: &str, dataset: &Dataset) method changed to from_query(query: &str, dataset: &Dataset, dialect: Option<Dialect>).
+
 ## [0.7.3] - 2024-01-04
 ### Removed
 - Dependency to matplotlib removed